pyproject.toml
[tool.poetry]
name = "tracely"
version = "0.1.0"
description = "Tracing tools"
authors = ["Emeli Dral <emeli.dral@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "tracely", from = "tracely/src"}]

[tool.poetry.dependencies]
python = "^3.8"
opentelemetry-api = ">=1.25.0"
opentelemetry-sdk = ">=1.25.0"
opentelemetry-proto = ">=1.25.0"
opentelemetry-exporter-otlp-proto-grpc = ">=1.25.0"
opentelemetry-exporter-otlp-proto-http = ">=1.25.0"
requests = ">=2.32.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
ruff = "^0.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4

[tool.flakje]
